requirements-render.txt
asgiref==3.10.0
bcrypt==5.0.0
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
dj-database-url==3.0.1
Django==4.2.27
django-cors-headers==4.9.0
django-filter==24.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
openpyxl==3.10.10
packaging==25.0
pillow==12.0.0
psycopg==3.3.2
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dotenv==1.2.1
reportlab==4.4.4
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
whitenoise==6.11.0
